requirements.txt
openai
wikipedia
python-dotenv
